Change [u]int to [u]int32 in winapi and adjust walk as required.

+++ src/pkg/walk/application.go
@@ -66,7 +66,7 @@ func (app *Application) SetSettings(value Settings) {
 func (app *Application) Exit(exitCode int) {
 	app.exiting = true
 	app.exitCode = exitCode
-	PostQuitMessage(exitCode)
+	PostQuitMessage(int32(exitCode))
 }
 
 func (app *Application) ExitCode() int {
